requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
redis>=4.0.0
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
SQLAlchemy>=1.4.0
psycopg2-binary>=2.9.0
PyYAML>=6.0
pytz>=2021.3
requests>=2.26.0
python-telegram-bot>=13.7
Werkzeug>=2.0.0
gunicorn>=20.1.0
python-dateutil>=2.8.2
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
nltk>=3.6.0
transformers>=4.0.0
torch>=1.9.0 